+++ matrix_lib/src/s21_eq_matrix.c
@@ -6,7 +6,7 @@ int s21_eq_matrix(matrix_t *A, matrix_t *B)
 		return FAILURE;
 	}
 	for (int i = 0; i < A->rows; i++) {
-		if (memcmp(A->matrix[i], B->matrix[i], A->columns * sizeof(double))) {
+		if (memcmp(A->matrix[i], B->matrix[i], A->columns * sizeof(float))) {
 			return FAILURE;
 		}
 	}
